+++ CHANGELOG.rst
@@ -4,8 +4,8 @@ Changelog
 0.5.0 (in development)
 ----------------------
 
-* Update `api.utils.create_read_depth_tsv` method to automatically detect `chr` string.
-* Add `sdk.utils.parse_input_bams` method.
+* Update ``api.utils.create_read_depth_tsv`` method to automatically detect ``chr`` string in input BAM.
+* Add ``sdk.utils.parse_input_bams`` method.
 
 
 0.4.1 (2021-09-21)